// src/app/model/averageStats.model.ts
export interface AverageStats {
    averagePoints: number;
    averageAssist: number;
    averageORebound: number;
    averageDRebound: number;
    averageSteals: number;
    averageTurnovers: number;
    averageBlocks: number;
    average2PM: number;
    average2PA: number;
    average3PM: number;
    average3PA: number;
    averageFTM: number;
    averageFTA: number;
    
    
    
    
}